renamed userProfile_model.py to user_profile_model.py
